+++ test/analysis_options_test.dart
@@ -39,6 +39,7 @@ analyzer:
 formatter:
   unknown_key: true
   page_width: 123
+  trailing_commas: preserve
 ''', custom: '');
     expect(json.decode(content), {
       'analyzer': {
@@ -51,6 +52,7 @@ formatter:
       },
       'formatter': {
         'page_width': 123,
+        'trailing_commas': 'preserve',
       },
     });
   });