+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.22.21
+
+- Preserve `formatter:trailing_commas` in `analysis_options.yaml` when doing `dart format`.
+
 ## 0.22.20
 
 - Reduced minimum description length to 50 characters.